clean up lambda tests
