+++ localstack/testing/aws/lambda_utils.py
@@ -1,8 +1,11 @@
 import json
+import os
 from typing import Literal
 
 from localstack.utils.common import to_str
 from localstack.utils.generic.wait_utils import ShortCircuitWaitException
+from localstack.utils.sync import retry
+from localstack.utils.testutil import get_lambda_log_events
 
 
 def update_done(client, function_name):
@@ -43,3 +46,73 @@ def get_invoke_init_type(
     return json.loads(to_str(invoke_result["Payload"].read()))["env"][
         "AWS_LAMBDA_INITIALIZATION_TYPE"
     ]
+
+
+lambda_role = {
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Effect": "Allow",
+            "Principal": {"Service": "lambda.amazonaws.com"},
+            "Action": "sts:AssumeRole",
+        }
+    ],
+}
+s3_lambda_permission = {
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Effect": "Allow",
+            "Action": [
+                "sqs:*",
+                "dynamodb:DescribeStream",
+                "dynamodb:GetRecords",
+                "dynamodb:GetShardIterator",
+                "dynamodb:ListStreams",
+                "kinesis:DescribeStream",
+                "kinesis:DescribeStreamSummary",
+                "kinesis:GetRecords",
+                "kinesis:GetShardIterator",
+                "kinesis:ListShards",
+                "kinesis:ListStreams",
+                "kinesis:SubscribeToShard",
+                "logs:CreateLogGroup",
+                "logs:CreateLogStream",
+                "logs:PutLogEvents",
+            ],
+            "Resource": ["*"],
+        }
+    ],
+}
+
+
+def _await_event_source_mapping_enabled(lambda_client, uuid, retries=30):
+    def assert_mapping_enabled():
+        assert lambda_client.get_event_source_mapping(UUID=uuid)["State"] == "Enabled"
+
+    retry(assert_mapping_enabled, sleep_before=2, retries=retries)
+
+
+def _await_dynamodb_table_active(dynamodb_client, table_name, retries=6):
+    def assert_table_active():
+        assert (
+            dynamodb_client.describe_table(TableName=table_name)["Table"]["TableStatus"] == "ACTIVE"
+        )
+
+    retry(assert_table_active, retries=retries, sleep_before=2)
+
+
+def _get_lambda_invocation_events(logs_client, function_name, expected_num_events, retries=30):
+    def get_events():
+        events = get_lambda_log_events(function_name, logs_client=logs_client)
+        assert len(events) == expected_num_events
+        return events
+
+    return retry(get_events, retries=retries, sleep_before=2)
+
+
+def is_old_provider():
+    return (
+        os.environ.get("TEST_TARGET") != "AWS_CLOUD"
+        and os.environ.get("PROVIDER_OVERRIDE_LAMBDA") != "asf"
+    )
